Add uprn methods to AddressManager
This exposes a named tuple with the fields that map on to the
ResidentialAddress model. The information for this is derived from
the Address model and the UprnToCouncil model.
Since the ResidentialAddress model is going to be removed this will
minimize changes to code which expected ResidentialAddress objects
to be returned.

Broadly:
 - `Address.objects.uprns_for_postcode(` corresponds to
   `ResidentialAddress.objects.filter(postcode=`.

 - `Address.objects.get_uprn(` corresponds to
   `ResidentialAddress.objects.get(`.