Add DistrictSet class to data_importers app
This will let us handle councils who provide polling districts as
polygons. We associate each polling district with a polling station,
then associate the station with all the UPRNs that fall within the
corresponding district.

Stations are joined to Districts based on ids in the data provided by
the council, as there is no constraint on a polling station falling
within it's polling district.

Addresses which fall on the boundary of a district, or within an
area created by overlapping districts are not assigned a polling
station. The latter case is the result of issues in digitisation and/or
simplification of the polygons on the council's end.