Add base classes for imports with districs
These are more or less the same as those in data_collection.
The main difference is the new `update_uprn_to_council_model()`
method of the DistrictSet class - which is used to kick off updating
the uprn table by doing the point in polygon lookup. The important
thing here is that the districts have to be in the db before this
will work.